feat: add students to project page
